refactor: cut_ages()
Removes the C implementation in favour of an R implementation
of `cut_ages()`. This refactoring has resulted in minor tweaks
to the given error messages as well as a small bug fix
relating to how `max_upper` argument was being rounded
(i.e. we were calling `round()` rather than raising to the
next integer(ish) number with `ceiling()` as intended).

Technically another breaking change but unlikely to cause
any issues in reality.